de-final-project-main/requirements.txt
tweepy==4.12.1
scikit-learn==0.24.2
kafka-python==1.4.6
pymongo==3.5.1
nltk==3.8
textblob==0.9.0
matplotlib==3.5.3